@@ -15,7 +15,7 @@
         "issues": "http://github.com/Innmind/Socket/issues"
     },
     "require": {
-        "php": "~7.2",
+        "php": "~7.4",
         "innmind/stream": "^1.5",
         "innmind/immutable": "^2.10",
         "innmind/event-bus": "^3.0",
